fix(test): move test to defineComponent.test-d.tsx
